requirements.txt
# Core Dependencies
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
plotly==5.18.0
flask==3.0.0
flask-cors==4.0.0
anthropic==0.25.0

# Data Processing
scipy==1.11.4
scikit-learn==1.4.0

# Visualization
seaborn==0.13.1
matplotlib==3.8.2

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
